Fix cleanup of jsdom environment
