+++ src/test/scala/outwatch/JSDomSpec.scala
@@ -1,0 +1,16 @@
+package outwatch
+
+
+import org.scalajs.dom._
+import org.scalatest.BeforeAndAfterEach
+
+
+abstract class JSDomSpec extends UnitSpec with BeforeAndAfterEach {
+  override def beforeEach(): Unit = {
+    document.body.innerHTML = ""
+
+    val root = document.createElement("div")
+    root.id = "app"
+    document.body.appendChild(root)
+  }
+}
